requirements.txt
aiogram==2.18
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
frozenlist==1.2.0
idna==3.3
lxml==4.7.1
multidict==5.2.0
mysql-connector-python==8.0.27
protobuf==3.19.1
pytz==2021.3
requests==2.26.0
soupsieve==2.3.1
urllib3==1.26.7
yarl==1.7.2
